+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 3.0.0 (IN PROGRESS)
 
+### Breaking changes
+
+- Legacy Calendar was replaced by Big Calendar.
+
 ### Features / Enhancements
 
 - Add keeping refresh dashboard on time range change (#167)
 - Add plugin e2e tests and remove cypress (#168, #170)
+- Remove Legacy calendar replaced by Big Calendar (#172)
 
 ## 2.5.0 (2024-03-20)